backend/requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pydantic-settings==2.5.0

# Image processing
Pillow==10.4.0
pillow-heif>=0.16.0
numpy>=1.26.4
scipy>=1.14.0
opencv-python-headless==4.10.0.84
imagehash==4.3.2

# AI / ML
google-generativeai==0.8.0
torch>=2.5.0
torchvision>=0.20.0
open-clip-torch>=3.3.0
timm>=1.0.0

# Watermark
invisible-watermark>=0.2.0

# Signal processing
PyWavelets>=1.6.0

# C2PA
c2pa-python==0.30.0

# Database
sqlalchemy[asyncio]==2.0.35
aiosqlite==0.20.0

# EXIF
exifread==3.0.0
